+++ core/commands/commands_test.go
@@ -165,6 +165,7 @@ func TestCommands(t *testing.T) {
 		"/ping",
 		"/provide",
 		"/provide/clear",
+		"/provide/stat",
 		"/pubsub",
 		"/pubsub/ls",
 		"/pubsub/peers",
