+++ docs/changelogs/v0.37.md
@@ -13,6 +13,7 @@ This release  was brought to you by the [Interplanetary Shipyard](https://ipship
   - [Clear provide queue when reprovide strategy changes](#clear-provide-queue-when-reprovide-strategy-changes)
   - [Named pins in `ipfs add` command](#-named-pins-in-ipfs-add-command)
   - [Removed unnecessary dependencies](#removed-unnecessary-dependencies)
+  - [Deprecated `ipfs stats reprovide`](#deprecated-ipfs-stats-reprovide)
 - [📦️ Important dependency updates](#-important-dependency-updates)
 - [📝 Changelog](#-changelog)
 - [👨‍👩‍👧‍👦 Contributors](#-contributors)
@@ -54,6 +55,14 @@ Kubo has been cleaned up by removing unnecessary dependencies and packages:
 - Removed `goprocess` dependency (replaced with native Go `context` patterns)
 
 These changes reduce the dependency footprint while improving code maintainability and following Go best practices.
+
+#### Deprecated `ipfs stats reprovide`
+
+The `ipfs stats reprovide` command has moved to `ipfs provide stat`. This was done to organize provider commands in one location.
+
+> [!NOTE]
+> `ipfs stats reprovide` still works, but is marked as deprecated and will be removed in a future release.
+
 
 #### 📦️ Important dependency updates
